+++ src/File/Content/OfStream.php
@@ -8,10 +8,7 @@ use Innmind\Filesystem\{
     Stream\LazyStream,
     Exception\FailedToLoadFile,
 };
-use Innmind\Stream\{
-    Readable,
-    PositionNotSeekable,
-};
+use Innmind\Stream\Readable;
 use Innmind\Immutable\{
     Sequence,
     SideEffect,
@@ -23,7 +20,7 @@ use Innmind\Immutable\{
 /**
  * @psalm-immutable
  */
-final class OfStream implements Content
+final class OfStream implements Content, Chunkable
 {
     /** @var callable(): Readable */
     private $load;
@@ -88,7 +85,7 @@ final class OfStream implements Content
 
     public function size(): Maybe
     {
-        return $this->stream()->size();
+        return $this->load()->size();
     }
 
     public function toString(): string
@@ -100,27 +97,37 @@ final class OfStream implements Content
         return Str::of('')->join($lines)->toString();
     }
 
+    public function chunks(): Sequence
+    {
+        return $this->sequence(static fn(Readable $stream) => $stream->read(8192));
+    }
+
     /**
      * This should be used only for reading chunk by chunk to persist the file
      * to the filesystem
      *
      * The stream returned MUST never be closed
      *
+     * @deprecated
      * @internal
      */
     public function stream(): Readable
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return ($this->load)();
+        return $this->load();
     }
 
     /**
+     * @param ?callable(Readable): Maybe<Str> $read
+     *
      * @return Sequence<Str>
      */
-    private function sequence(): Sequence
+    private function sequence(callable $read = null): Sequence
     {
-        return Sequence::lazy(function($cleanup) {
-            $stream = $this->stream();
+        /** @var callable(Readable): Maybe<Str> */
+        $read ??= static fn(Readable $stream): Maybe => $stream->readLine();
+
+        return Sequence::lazy(function($cleanup) use ($read) {
+            $stream = $this->load();
             $rewind = static function() use ($stream): void {
                 $_ = $stream->rewind()->match(
                     static fn() => null, // rewind successfull
@@ -138,15 +145,19 @@ final class OfStream implements Content
                 // we yield an empty line when the readLine() call doesn't return
                 // anything otherwise it will fail to load empty files or files
                 // ending with the "end of line" character
-                yield $stream
-                    ->readLine()
-                    ->match(
-                        static fn($line) => $line,
-                        static fn() => Str::of(''),
-                    );
+                yield $read($stream)->match(
+                    static fn($line) => $line,
+                    static fn() => Str::of(''),
+                );
             }
 
             $rewind();
         });
+    }
+
+    private function load(): Readable
+    {
+        /** @psalm-suppress ImpureFunctionCall */
+        return ($this->load)();
     }
 }